Time: 100 ms (60.88%), Space: 65.2 MB (8.84%) - LeetHub

+++ 0295-find-median-from-data-stream/0295-find-median-from-data-stream.java
@@ -6,7 +6,7 @@ class MedianFinder {
     }
     
     public void addNum(int num) {
-        if(maxHeap.isEmpty() || (long)num <= maxHeap.peek()){
+        if(maxHeap.isEmpty() || (double) num <= maxHeap.peek()){
             maxHeap.add((double)num);
         }else{
             minHeap.add((double)num);
